pyproject.toml
[project]
name = "gemma-earth"
version = "0.1.0"
description = "LoRA fine-tuning pipeline for Gemma 3 4B IT on EarthDial for remote sensing scene classification."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
 "datasets>=2.14.4",
 "flax>=0.12.6",
 "gcsfs>=2026.3.0",
 "google-tunix[prod]>=0.1.6",
 "huggingface-hub>=0.36.2",
 "jax[tpu]>=0.9.2",
 "matplotlib>=3.10.8",
 "pillow>=12.2.0",
 "pyarrow<21",
 "pydantic-settings>=2.13.1",
 "python-dotenv>=1.2.2",
 "qwix",
 "rich>=14.3.3",
 "rich-pixels>=3.0.1",
 "setuptools>=82.0.1",
 "tensorboard>=2.20.0",
 "tpu-info>=0.10.0",
]

[project.scripts]
gemma-earth = "main:main"

[build-system]
requires = ["uv_build>=0.11.3,<0.12.0"]
build-backend = "uv_build"

[tool.uv.sources]
qwix = { git = "https://github.com/google/qwix" }
